Move lib/types.ts to src/types/basic.ts
