+++ src/compose/service.ts
@@ -23,7 +23,7 @@ import { sanitiseComposeConfig } from './sanitise';
 import { getPathOnHost } from '../lib/fs-utils';
 
 import log from '../lib/supervisor-console';
-import { EnvVarObject } from '../lib/types';
+import { EnvVarObject } from '../types';
 
 const SERVICE_NETWORK_MODE_REGEX = /service:\s*(.+)/;
 const CONTAINER_NETWORK_MODE_REGEX = /container:\s*(.+)/;
