+++ src/types/state.ts
@@ -1,7 +1,9 @@
+// TODO: move all these exported types to ../compose/types
 import { ComposeNetworkConfig } from '../compose/types/network';
 import { ServiceComposeConfig } from '../compose/types/service';
 import { ComposeVolumeConfig } from '../compose/volume';
-import { EnvVarObject, LabelObject } from '../lib/types';
+
+import { EnvVarObject, LabelObject } from './basic';
 
 import App from '../compose/app';
